feat: Add quest Bloodfire Ritual (#50)
This is a more complex quest, with more of a feast or famine dynamic. We don't want too many of those but I think this fits as a challenge quest, and I'm confident it has the potential to be fun and reasonably balanced. The game will be harder for long enough that the strong card you get from completing the quest isn't a solo game winner (since you'll be facing act 3 and 4 by the time you get it).

This quest is big enough that it's harder to know what the right numbers are for all the different parts of it. I expect it will need some tweaking once people are able to playtest, but that's okay.

* Bloodfire Ritual quest: add quest
* Bloodfire Ritual: simplify the quest reward
* Bloodfire Ritual quest: use makeContributionPath instead of makeImagePath (and update other quests of mine as well)